+++ backend/files/sync.go
@@ -15,13 +15,16 @@ func (idx *Index) UpdateMetadata(info *FileInfo) bool {
 	return true
 }
 
-// GetMetadataInfo retrieves the FileInfo from the specified directory in the index.
+// GetMetadataInfo retrieves the FileInfo from the specified file or directory in the index.
 func (idx *Index) GetReducedMetadata(target string, isDir bool) (*FileInfo, bool) {
 	idx.mu.Lock()
 	defer idx.mu.Unlock()
-	checkDir := idx.makeIndexPath(target)
+	checkDir := idx.MakeIndexPath(target)
 	if !isDir {
-		checkDir = idx.makeIndexPath(filepath.Dir(target))
+		checkDir = idx.MakeIndexPath(filepath.Dir(target))
+	}
+	if checkDir == "" {
+		checkDir = "/"
 	}
 	dir, exists := idx.Directories[checkDir]
 	if !exists {
@@ -48,13 +51,16 @@ func (idx *Index) GetReducedMetadata(target string, isDir bool) (*FileInfo, bool
 
 }
 
-// GetMetadataInfo retrieves the FileInfo from the specified directory in the index.
+// raw directory info retrieval -- does not work on files, only returns a directory
 func (idx *Index) GetMetadataInfo(target string, isDir bool) (*FileInfo, bool) {
 	idx.mu.RLock()
 	defer idx.mu.RUnlock()
-	checkDir := idx.makeIndexPath(target)
+	checkDir := idx.MakeIndexPath(target)
 	if !isDir {
-		checkDir = idx.makeIndexPath(filepath.Dir(target))
+		checkDir = idx.MakeIndexPath(filepath.Dir(target))
+	}
+	if checkDir == "" {
+		checkDir = "/"
 	}
 	dir, exists := idx.Directories[checkDir]
 	return dir, exists