+++ backend/http/search.go
@@ -3,6 +3,8 @@ package http
 import (
 	"fmt"
 	"net/http"
+	"net/url"
+	"path/filepath"
 	"strings"
 
 	"github.com/gtsteffaniak/filebrowser/backend/files"
@@ -59,8 +61,12 @@ func searchHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (i
 	if source == "" {
 		source = config.Server.DefaultSource.Name
 	}
-	searchScope := strings.TrimPrefix(r.URL.Query().Get("scope"), ".")
-	searchScope = strings.TrimPrefix(searchScope, "/")
+	scope := r.URL.Query().Get("scope")
+	unencodedScope, err := url.QueryUnescape(scope)
+	if err != nil {
+		return http.StatusBadRequest, fmt.Errorf("invalid path encoding: %v", err)
+	}
+	searchScope := strings.TrimPrefix(unencodedScope, ".")
 	// Retrieve the User-Agent and X-Auth headers from the request
 	sessionId := r.Header.Get("SessionId")
 	index := files.GetIndex(source)
@@ -71,11 +77,9 @@ func searchHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (i
 	if err != nil {
 		return http.StatusForbidden, err
 	}
-	userScope := strings.TrimPrefix(userscope, ".")
-	combinedScope := strings.TrimPrefix(userScope+"/"+searchScope, "/")
-
+	combinedPath := index.MakeIndexPath(filepath.Join(userscope, searchScope))
 	// Perform the search using the provided query and user scope
-	response := index.Search(query, combinedScope, sessionId)
+	response := index.Search(query, combinedPath, sessionId)
 	// Set the Content-Type header to application/json
 	w.Header().Set("Content-Type", "application/json")
 	return renderJSON(w, r, response)