+++ frontend/src/components/Breadcrumbs.vue
@@ -1,17 +1,16 @@
 <template>
   <div class="breadcrumbs">
-    <component
-      :is="element"
-      :to="homePath"
-      :aria-label="$t('files.home')"
-      :title="$t('files.home')"
-    >
+    <router-link :to="base" :aria-label="$t('files.home')" :title="$t('files.home')">
       <i class="material-icons">home</i>
-    </component>
+    </router-link>
 
-    <span v-for="(link, index) in items" :key="index">
+    <span
+      :aria-label="'breadcrumb-link-' + link.name"
+      v-for="(link, index) in items"
+      :key="index"
+    >
       <span class="chevron"><i class="material-icons">keyboard_arrow_right</i></span>
-      <component :is="element" :to="link.url">{{ link.name }}</component>
+      <router-link :to="link.url">{{ link.name }}</router-link>
     </span>
     <action style="display: contents" v-if="showShare" icon="share" show="share" />
     <div v-if="isCardView">
@@ -32,6 +31,7 @@
 
 <script>
 import { state, mutations, getters } from "@/store";
+import { baseURL } from "@/utils/constants.js";
 import { extractSourceFromPath, removePrefix, removeLeadingSlash } from "@/utils/url.js";
 import Action from "@/components/Action.vue";
 
@@ -43,11 +43,11 @@ export default {
   data() {
     return {
       gallerySize: state.user.gallerySize,
-      homePath: "/files/",
+      base: "/files/",
       path: "",
     };
   },
-  props: ["base", "noLink"],
+  props: ["noLink"],
   mounted() {
     this.updatePaths();
   },
@@ -67,37 +67,23 @@ export default {
       }
 
       let breadcrumbs = [];
-
-      for (let i = 0; i < parts.length; i++) {
-        if (i === 0) {
-          breadcrumbs.push({
-            name: decodeURIComponent(parts[i]),
-            url: this.base + "/" + parts[i] + "/",
-          });
-        } else {
-          breadcrumbs.push({
-            name: decodeURIComponent(parts[i]),
-            url: breadcrumbs[i - 1].url + parts[i] + "/",
-          });
-        }
-      }
+      let buildRef = this.base;
+      parts.forEach((element) => {
+        buildRef = buildRef + encodeURIComponent(element) + "/";
+        breadcrumbs.push({
+          name: decodeURIComponent(element),
+          url: buildRef,
+        });
+      });
 
       if (breadcrumbs.length > 3) {
         while (breadcrumbs.length !== 4) {
           breadcrumbs.shift();
         }
-
         breadcrumbs[0].name = "...";
       }
 
       return breadcrumbs;
-    },
-    element() {
-      if (this.noLink !== undefined) {
-        return "span";
-      }
-
-      return "router-link";
     },
     showShare() {
       return (
@@ -111,13 +97,15 @@ export default {
   methods: {
     updatePaths() {
       const result = extractSourceFromPath(getters.routePath());
-      this.path = removePrefix(result.path, "files");
       if (getters.currentView() == "share") {
-        this.homePath = getters.sharePathBase();
-        this.path = removePrefix(getters.routePath(), this.homePath + "/");
+        this.base = getters.sharePathBase();
+        this.path = removePrefix(getters.routePath(), this.base + "/");
+      } else {
+        this.path = decodeURIComponent(result.path);
       }
-      if (state.sources.count > 1) {
-        this.homePath = "/files/" + result.source;
+      this.base = baseURL;
+      if (state.serverHasMultipleSources) {
+        this.base = `${this.base}${result.source}/`;
       }
     },
     updateGallerySize(event) {