+++ frontend/src/views/Files.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <breadcrumbs v-if="showBreadCrumbs" base="/files" />
+    <breadcrumbs v-if="showBreadCrumbs" />
     <errors v-if="error" :errorCode="error.status" />
     <component v-else-if="currentViewLoaded" :is="currentView"></component>
     <div v-else>
@@ -103,7 +103,7 @@ export default {
     async fetchData() {
       // lets redirect if multiple sources and user went to /files/
       if (
-        state.sources.count > 1 &&
+        state.serverHasMultipleSources &&
         (getters.routePath() === "/files/" || getters.routePath() === "/files")
       ) {
         router.push(`/files/${state.sources.current}`);